+++ tests/integration/redis_tests.py
@@ -78,67 +78,6 @@ class RedisIntegrationTests(unittest.TestCase):
         self.assertIsNone(span_observer.on_finish_exc_info)
 
 
-class RedisClusterIntegrationTests(unittest.TestCase):
-    def setUp(self):
-        self.baseplate_observer = TestBaseplateObserver()
-
-        baseplate = Baseplate({"rediscluster.url": f"redis://{redis_endpoint}/0"})
-        baseplate.register(self.baseplate_observer)
-        baseplate.configure_context({"rediscluster": RedisClient()})
-
-        self.context = baseplate.make_context_object()
-        self.server_span = baseplate.make_server_span(self.context, "test")
-
-    def test_simple_command(self):
-        with self.server_span:
-            result = self.context.rediscluster.ping()
-
-        self.assertTrue(result)
-
-        server_span_observer = self.baseplate_observer.get_only_child()
-        span_observer = server_span_observer.get_only_child()
-        self.assertEqual(span_observer.span.name, "rediscluster.PING")
-        self.assertTrue(span_observer.on_start_called)
-        self.assertTrue(span_observer.on_finish_called)
-        self.assertIsNone(span_observer.on_finish_exc_info)
-
-    def test_error(self):
-        with self.server_span:
-            with self.assertRaises(redis.ResponseError):
-                self.context.rediscluster.execute_command("crazycommand")
-
-        server_span_observer = self.baseplate_observer.get_only_child()
-        span_observer = server_span_observer.get_only_child()
-        self.assertTrue(span_observer.on_start_called)
-        self.assertTrue(span_observer.on_finish_called)
-        self.assertIsNotNone(span_observer.on_finish_exc_info)
-
-    def test_lock(self):
-        with self.server_span:
-            with self.context.rediscluster.lock("foo"):
-                pass
-
-        server_span_observer = self.baseplate_observer.get_only_child()
-
-        self.assertGreater(len(server_span_observer.children), 0)
-        for span_observer in server_span_observer.children:
-            self.assertTrue(span_observer.on_start_called)
-            self.assertTrue(span_observer.on_finish_called)
-
-    def test_pipeline(self):
-        with self.server_span:
-            with self.context.rediscluster.pipeline("foo") as pipeline:
-                pipeline.ping()
-                pipeline.execute()
-
-        server_span_observer = self.baseplate_observer.get_only_child()
-        span_observer = server_span_observer.get_only_child()
-        self.assertEqual(span_observer.span.name, "rediscluster.pipeline_foo")
-        self.assertTrue(span_observer.on_start_called)
-        self.assertTrue(span_observer.on_finish_called)
-        self.assertIsNone(span_observer.on_finish_exc_info)
-
-
 class RedisMessageQueueTests(unittest.TestCase):
     qname = "redisTestQueue"
